+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased 
+
+#### Dependencies 
+- Kotlin updated to 1.4.0
+- kotlinx-coroutines updated to 1.3.9
+- Klock updated to 1.12.0
+
+#### Meta
+- Gradle wrapper updated to 6.6
+
 ## 0.4.0 (2020-07-25)
 
 #### Added